requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
waitress==3.0.2
